+++ DesignerMacros/src/services-cqrs-crud/_common/ensureHasField.ts
@@ -26,8 +26,10 @@ function ensureHasField(options: {
         field.setOrder(order);
     }
 
+    console.warn(field.getName());
     if (mappingSettingsId != null) {
         field.setMapping(fieldDetail.mappingPath, mappingSettingsId);
+        console.warn("mapping:" + fieldDetail.mappingPath);
     }
 
     return field;