+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpConstructor.cs
@@ -8,23 +8,32 @@ namespace Intent.Modules.Common.CSharp.Builder;
 
 public class CSharpConstructor : CSharpMember<CSharpConstructor>, IHasCSharpStatements, IHasICSharpParameters, ICSharpReferenceable
 {
+    // This class will be used for "static", "instance" and "primary" constructors.
+    //https://learn.microsoft.com/en-us/dotnet/csharp/programming-guide/classes-and-structs/instance-constructors
+    
     public CSharpClass Class { get; }
     public string AccessModifier { get; private set; } = "public ";
     public CSharpConstructorCall ConstructorCall { get; private set; }
     public IList<CSharpConstructorParameter> Parameters { get; } = new List<CSharpConstructorParameter>();
     public List<CSharpStatement> Statements { get; } = new();
     public string Name => Class.Name;
+    public bool IsPrimaryConstructor { get; private set; }
 
     IList<CSharpStatement> IHasCSharpStatements.Statements => this.Statements;
     IEnumerable<ICSharpParameter> IHasICSharpParameters.Parameters => this.Parameters;
 
-    public CSharpConstructor(CSharpClass @class)
+    public CSharpConstructor(CSharpClass @class) : this(@class, false)
+    {
+    }
+
+    public CSharpConstructor(CSharpClass @class, bool isPrimaryConstructor)
     {
         BeforeSeparator = CSharpCodeSeparatorType.EmptyLines;
         AfterSeparator = CSharpCodeSeparatorType.EmptyLines;
         Parent = @class;
         Class = @class;
         File = @class.File;
+        IsPrimaryConstructor = isPrimaryConstructor;
     }
 
     /// <summary>
@@ -77,6 +86,16 @@ public class CSharpConstructor : CSharpMember<CSharpConstructor>, IHasCSharpStat
         var param = new CSharpConstructorParameter(type, name, this);
         Parameters.Add(param);
         configure?.Invoke(param);
+        IntroduceBackingMembersIfForPrimaryConstructor(param);
+        return this;
+    }
+
+    public CSharpConstructor InsertParameter(int index, string type, string name, Action<CSharpConstructorParameter> configure = null)
+    {
+        var param = new CSharpConstructorParameter(type, name, this);
+        Parameters.Insert(index, param);
+        configure?.Invoke(param);
+        IntroduceBackingMembersIfForPrimaryConstructor(param);
         return this;
     }
 
@@ -147,21 +166,55 @@ public class CSharpConstructor : CSharpMember<CSharpConstructor>, IHasCSharpStat
 
     public CSharpConstructor CallsThis(Action<CSharpConstructorCall> configure = null)
     {
+        if (IsPrimaryConstructor)
+        {
+            throw new InvalidOperationException($"Cannot supply 'this' constructor call to a primary constructor: {Class.TypeDefinitionType} {Class.Name}.");
+        }
         ConstructorCall = CSharpConstructorCall.This();
         configure?.Invoke(ConstructorCall);
         return this;
     }
-
+    
     public override string GetText(string indentation)
     {
+        if (IsPrimaryConstructor)
+        {
+            return Parameters.Count == 0 ? string.Empty : $"({ToStringParameters(indentation)})";
+        }
+        
         return $@"{GetComments(indentation)}{GetAttributes(indentation)}{indentation}{AccessModifier}{Class.Name}({ToStringParameters(indentation)}){ConstructorCall?.ToString() ?? string.Empty}
 {indentation}{{{Statements.ConcatCode($"{indentation}    ")}
 {indentation}}}";
     }
 
+    private void IntroduceBackingMembersIfForPrimaryConstructor(CSharpConstructorParameter param)
+    {
+        if (!IsPrimaryConstructor)
+        {
+            return;
+        }
+
+        switch (Class.TypeDefinitionType)
+        {
+            case CSharpClass.Type.Class:
+                var field = CSharpField.CreateFieldOmittedFromRender(param.Type, param.Name, null);
+                Class.Fields.Add(field);
+                break;
+            case CSharpClass.Type.Record:
+                var property = CSharpProperty.CreatePropertyOmittedFromRender(param.Type, param.Name, Class);
+                Class.Properties.Add(property);
+                break;
+        }
+    }
+
     private string ToStringParameters(string indentation)
     {
-        if (Parameters.Count > 1 && $"{indentation}{AccessModifier}{Class.Name}(".Length + Parameters.Sum(x => x.ToString().Length) > 120)
+        // This length is more or less the same for an instance and primary ctor declaration on a single line
+        var estimatedLength = $"{indentation}{AccessModifier}{Class.Name}(".Length + Parameters.Sum(x => x.ToString().Length);
+        
+        const int maxLineLength = 120;
+        
+        if (Parameters.Count > 1 && estimatedLength > maxLineLength)
         {
             return string.Join($@",
 {indentation}    ", Parameters.Select(x => x.ToString()));
@@ -171,5 +224,4 @@ public class CSharpConstructor : CSharpMember<CSharpConstructor>, IHasCSharpStat
             return string.Join(", ", Parameters.Select(x => x.ToString()));
         }
     }
-
 }