+++ Modules/Intent.Modules.Common.CSharp/Mapping/Mappings/MethodInvocationMapping.cs
@@ -5,6 +5,7 @@ using System.Reflection.Metadata;
 using Intent.Metadata.Models;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.TypeResolution;
 
 namespace Intent.Modules.Common.CSharp.Mapping;
@@ -30,16 +31,18 @@ public class MethodInvocationMapping : CSharpMappingBase
 
     public override CSharpStatement GetSourceStatement()
     {
-        var invocation = new CSharpInvocationStatement(GetTargetPathExpression());
+		var invocation = new CSharpInvocationStatement(GetTargetPathExpression());
 
 		var typeTemplate = _template.GetTypeInfo(((IElement)Model).ParentElement.AsTypeReference())?.Template as ICSharpFileBuilderTemplate;
 		// Determine if this model is a method on the class:
 		if (typeTemplate?.CSharpFile.Classes.FirstOrDefault()?.TryGetReferenceForModel(Model.Id, out var reference) == true && reference is CSharpClassMethod method)
         {
+			//Link the method call so the builder can work out the Async invocation syntax
+			invocation.Invokes(method);
 			foreach (var parameter in method.Parameters)
 			{
 				bool optional = parameter.DefaultValue != null;
-				var mappedChild = Children.FirstOrDefault(c => string.Compare(c.Mapping.TargetElement.Name, parameter.Name, true) == 0);
+				var mappedChild = Children.FirstOrDefault(c => c.Model.Name.Equals(parameter.Name, StringComparison.InvariantCultureIgnoreCase));
 				if (mappedChild != null)
 				{
 					invocation.AddArgument(mappedChild.GetSourceStatement());