+++ Modules/Intent.Modules.Common.CSharp/Mapping/Mappings/ObjectInitializationMapping.cs
@@ -8,6 +8,7 @@ using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.TypeResolution;
+using System.Net.Http.Headers;
 
 namespace Intent.Modules.Common.CSharp.Mapping
 {
@@ -43,7 +44,10 @@ namespace Intent.Modules.Common.CSharp.Mapping
 
                 if (Model.TypeReference.IsCollection)
                 {
-                    Template.AddUsing("System.Linq");
+                    var m = new SelectToListMapping(_mappingModel, _template);
+                    m.Parent = this.Parent;
+					return m.GetSourceStatement();
+                    /*Template.AddUsing("System.Linq");
                     var chain = new CSharpMethodChainStatement($"{GetSourcePathText()}{(Mapping.SourceElement.TypeReference.IsNullable ? "?" : "")}").WithoutSemicolon();
                     var select = new CSharpInvocationStatement($"Select").WithoutSemicolon();
 
@@ -61,7 +65,7 @@ namespace Intent.Modules.Common.CSharp.Mapping
                     var init = chain
                         .AddChainStatement(select)
                         .AddChainStatement("ToList()");
-                    return init;
+                    return init;*/
                 }
                 else
                 {
@@ -71,8 +75,9 @@ namespace Intent.Modules.Common.CSharp.Mapping
                     }
                     else
                     {
-                        // TODO: add ternary check to mappings for when the source path could be nullable.
-                        var lastTargetPathElement = GetTargetPath().Last().Element;
+
+						// TODO: add ternary check to mappings for when the source path could be nullable.
+						var lastTargetPathElement = GetTargetPath().Last().Element;
                         SetTargetReplacement(lastTargetPathElement, null); // Needed for inheritance mappings - path element to be removed from invocation path
                         if (lastTargetPathElement.TypeReference.Element is not null)
                         {
@@ -176,24 +181,13 @@ namespace Intent.Modules.Common.CSharp.Mapping
             }
 
             var constructors = template.CSharpFile.TypeDeclarations.SelectMany(s => s.Constructors).ToArray();
-            var mapTargetElements = FindPropertyMappingsInHierarchy(Children).Select(s => s.Mapping.TargetElement).ToList();
+            var mapTargetElements = FindPropertyMappingsInHierarchy(Children).Where(s => s.Mapping != null).Select(s => s.Mapping.TargetElement).ToList();
 
             return constructors
                 .Where(ctor => mapTargetElements
                     .All(target => ctor.Parameters.Any(param => param.TryGetReferenceForModel(target, out var match) && param.Name == match.Name)))
                 .Select(s => (Ctor: s, MetadataElement: (IElement)returnTypeElement))
                 .ToList();
-        }
-
-        private string GetVariableNameForSelect()
-        {
-            var variableName = string.Join("", Model.Name.Where(char.IsUpper).Select(char.ToLower));
-            if (string.IsNullOrEmpty(variableName))
-            {
-                variableName = Char.ToLower(Model.Name[0]).ToString();
-            }
-
-            return variableName;
         }
 
         public override CSharpStatement GetTargetStatement()